add bcrypt; change logic for database reset; change len for password in database; add hash and check password func in utils;

+++ src/routers/database.py
@@ -1,14 +1,13 @@
-from fastapi import APIRouter, Depends, Cookie
-from typing import Annotated
-from src.database import engine
+from fastapi import APIRouter, HTTPException
+from src.database import engine, database_reset_key
 from src.models.database import Base
-from src.utils import admin_check
 
 router = APIRouter(prefix="/db", tags=["database"])
 
 @router.post("/reset", summary="[WARNING] DELETE AND RECREATE ALL TABLES IN DATABASE")
-def db_reset(token: Annotated[str, Cookie()]):
-    admin_check(token)
+def db_reset(key: str):
+    if database_reset_key != key:
+        raise HTTPException(403, "Key is invalid")
     Base.metadata.drop_all(engine)
     Base.metadata.create_all(engine)
     return {"status": "Database was reset."}
